@@ -30,86 +30,88 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.2.2",
-    "@angular/common": "^14.2.2",
-    "@angular/compiler": "^14.2.2",
-    "@angular/core": "^14.2.2",
-    "@angular/forms": "^14.2.2",
-    "@angular/platform-browser": "^14.2.2",
-    "@angular/platform-browser-dynamic": "^14.2.2",
-    "@angular/router": "^14.2.2",
+    "@angular/animations": "^15.1.1",
+    "@angular/common": "^15.1.1",
+    "@angular/compiler": "^15.1.1",
+    "@angular/core": "^15.1.1",
+    "@angular/forms": "^15.1.1",
+    "@angular/platform-browser": "^15.1.1",
+    "@angular/platform-browser-dynamic": "^15.1.1",
+    "@angular/router": "^15.1.1",
     "@ngneat/lib": "^5.0.0",
-    "@ngrx/effects": "^14.3.1",
-    "@ngrx/router-store": "^14.3.1",
-    "@ngrx/store": "^14.3.1",
-    "@reduxjs/toolkit": "~1.8.5",
-    "rxjs": "~7.5.6",
-    "tslib": "^2.4.0",
-    "zone.js": "~0.11.8"
+    "@ngrx/effects": "^15.1.0",
+    "@ngrx/router-store": "^15.1.0",
+    "@ngrx/store": "^15.1.0",
+    "@reduxjs/toolkit": "^1.9.1",
+    "rxjs": "~7.5.7",
+    "tslib": "^2.4.1",
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^14.0.1",
-    "@angular-devkit/build-angular": "^14.2.3",
-    "@angular-eslint/builder": "^14.1.1",
-    "@angular-eslint/eslint-plugin": "^14.1.1",
-    "@angular-eslint/eslint-plugin-template": "^14.1.1",
-    "@angular-eslint/schematics": "^14.1.1",
-    "@angular-eslint/template-parser": "^14.1.1",
-    "@angular/cli": "^14.2.3",
-    "@angular/compiler-cli": "^14.2.2",
-    "@commitlint/cli": "^17.1.2",
-    "@commitlint/config-conventional": "^17.1.0",
-    "@ngrx/store-devtools": "^14.3.1",
-    "@semantic-release/changelog": "^6.0.1",
+    "@angular-builders/jest": "^15.0.0",
+    "@angular-devkit/build-angular": "^15.1.2",
+    "@angular-eslint/builder": "^15.2.0",
+    "@angular-eslint/eslint-plugin": "^15.2.0",
+    "@angular-eslint/eslint-plugin-template": "^15.2.0",
+    "@angular-eslint/schematics": "^15.2.0",
+    "@angular-eslint/template-parser": "^15.2.0",
+    "@angular/cli": "^15.1.2",
+    "@angular/compiler-cli": "^15.1.1",
+    "@commitlint/cli": "^17.4.2",
+    "@commitlint/config-conventional": "^17.4.2",
+    "@ngrx/store-devtools": "^15.1.0",
+    "@semantic-release/changelog": "^6.0.2",
     "@semantic-release/git": "^10.0.1",
     "@tailwindcss/forms": "^0.5.3",
-    "@testing-library/angular": "^12.1.2",
+    "@testing-library/angular": "^13.1.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^29.0.3",
+    "@types/jest": "^29.2.6",
     "@types/node": "^12.20.55",
     "@types/node-fetch": "^2.6.2",
-    "@typescript-eslint/eslint-plugin": "^5.38.0",
-    "@typescript-eslint/parser": "^5.38.0",
-    "all-contributors-cli": "^6.20.4",
-    "angular-cli-ghpages": "1.0.3",
+    "@typescript-eslint/eslint-plugin": "^5.48.2",
+    "@typescript-eslint/parser": "^5.48.2",
+    "all-contributors-cli": "^6.24.0",
+    "angular-cli-ghpages": "1.0.5",
     "cpx": "^1.5.0",
     "cross-env": "^7.0.3",
-    "eslint": "^8.23.1",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.0.4",
-    "eslint-plugin-jest-dom": "^4.0.2",
-    "eslint-plugin-jsdoc": "^39.3.6",
+    "eslint": "^8.32.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.1",
+    "eslint-plugin-jest-dom": "^4.0.3",
+    "eslint-plugin-jsdoc": "^39.6.7",
     "eslint-plugin-ngrx": "^2.1.4",
     "eslint-plugin-prefer-arrow": "^1.2.3",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-rxjs": "^5.0.2",
-    "eslint-plugin-testing-library": "^5.6.4",
+    "eslint-plugin-tailwindcss": "^3.8.0",
+    "eslint-plugin-testing-library": "^5.9.1",
     "git-cz": "^4.9.0",
-    "husky": "^8.0.1",
-    "jest": "^29.0.3",
-    "jest-preset-angular": "^12.2.2",
+    "husky": "^8.0.3",
+    "jest": "^29.3.1",
+    "jest-preset-angular": "^12.2.5",
     "jsonc-parser": "^3.2.0",
-    "lint-staged": "^13.0.3",
-    "msw": "^0.47.3",
-    "ng-packagr": "^14.2.1",
-    "node-fetch": "^2.6.7",
-    "postcss": "^8.4.16",
-    "prettier": "^2.7.1",
+    "lint-staged": "^13.1.0",
+    "msw": "^0.49.3",
+    "ng-packagr": "^15.1.1",
+    "node-fetch": "^2.6.8",
+    "postcss": "^8.4.21",
+    "prettier": "^2.8.3",
     "prettier-plugin-organize-attributes": "0.0.5",
+    "prettier-plugin-tailwindcss": "^0.2.1",
     "protractor": "~7.0.0",
     "semantic-release": "^19.0.5",
-    "stylelint": "^14.12.0",
-    "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-recommended-scss": "^7.0.0",
-    "stylelint-config-standard": "^28.0.0",
-    "stylelint-config-standard-scss": "^5.0.0",
+    "stylelint": "^14.16.1",
+    "stylelint-config-prettier": "^9.0.4",
+    "stylelint-config-recommended-scss": "^8.0.0",
+    "stylelint-config-standard": "^29.0.0",
+    "stylelint-config-standard-scss": "^6.1.0",
     "stylelint-scss": "^4.3.0",
-    "tailwindcss": "^3.1.8",
-    "ts-jest": "^29.0.1",
+    "tailwindcss": "^3.2.4",
+    "ts-jest": "^29.0.5",
     "ts-node": "~10.9.1",
-    "typescript": "~4.8.3"
+    "typescript": "~4.8.4"
   },
   "config": {
     "commitizen": {
