@@ -1,2 +1,2 @@
 export * from './hooks-types';
-export * from './hooks-ts41-types';
+export * from './namedHooks';
