@@ -6,19 +6,19 @@ import { Character, CharacterStatus } from '../models';
   selector: 'app-character-card',
   template: `
     <div
-      class="flex flex-col items-center justify-center h-full p-4 transition-colors duration-200 bg-white rounded-lg shadow hover:bg-gray-100"
+      class="flex h-full flex-col items-center justify-center rounded-lg bg-white p-4 shadow transition-colors duration-200 hover:bg-gray-100"
     >
       <div
-        class="inline-flex w-40 h-40 overflow-hidden border border-gray-200 rounded-full shadow-lg dark:border-gray-800"
+        class="inline-flex h-40 w-40 overflow-hidden rounded-full border border-gray-200 shadow-lg dark:border-gray-800"
       >
-        <img class="w-full h-full" [src]="character.image" [alt]="character.name" />
+        <img class="h-full w-full" [src]="character.image" [alt]="character.name" />
       </div>
 
       <h2 class="mt-4 text-xl font-bold">{{ character.name }}</h2>
       <h6 class="mt-1 text-sm font-medium">{{ character.species }} - {{ character.gender }}</h6>
-      <div class="flex items-center mt-1 space-x-2">
+      <div class="mt-1 flex items-center space-x-2">
         <span
-          class="w-3 h-3 rounded-full shadow"
+          class="h-3 w-3 rounded-full shadow"
           [ngClass]="
             character.status.toLowerCase() === statusTypes.dead
               ? 'bg-red-400'
@@ -30,18 +30,18 @@ import { Character, CharacterStatus } from '../models';
         <h6 class="text-sm font-medium">{{ character.status }}</h6>
       </div>
 
-      <p class="self-start mt-3 text-xs text-gray-500 dark:text-gray-400">
+      <p class="mt-3 self-start text-xs text-gray-500 dark:text-gray-400">
         Last known location:
         <span class="inline-block text-indigo-700 hover:text-indigo-800">{{ character.location.name }}</span>
       </p>
       <div
         *ngIf="episodeQuery.state$ | async as episodeQuery"
-        class="self-start mt-3 text-xs text-gray-500 dark:text-gray-400"
+        class="mt-3 self-start text-xs text-gray-500 dark:text-gray-400"
       >
         First seen:
         <div
           *ngIf="episodeQuery.isLoading; else episodeName"
-          class="inline-block w-32 h-4 ml-1 bg-indigo-200 rounded animate-pulse"
+          class="ml-1 inline-block h-4 w-32 animate-pulse rounded bg-indigo-200"
         ></div>
         <ng-template #episodeName>
           <span class="inline-block text-indigo-700 hover:text-indigo-800">
